# src/main/resources/static/api-docs.yml
openapi: 3.0.0
info:
  title: "communication-service"
  version: 1.0.0
  description: >-
    Business management service.
  license:
    name: Apache-2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
paths:
  /email:
    post:
      tags:
        - Email
      description:
        Send one email
      operationId: sendOne
      security:
        - ServiceAuth: [ ]
      requestBody:
        required: true
        content:
          application/vnd.kaiqkt_email_v1+json:
            schema:
              $ref: '#/components/schemas/EmailV1'
      responses:
        204:
          description: Email sent with successfully
  /sms:
    post:
      tags:
        - Sms
      description:
        Send one sms
      operationId: sendOne
      security:
        - ServiceAuth: [ ]
      requestBody:
        required: true
        content:
          application/vnd.kaiqkt_sms_v1+json:
            schema:
              $ref: '#/components/schemas/SmsV1'
      responses:
        204:
          description: Sms sent with successfully
  /push:
    post:
      tags:
        - Push
      description:
        Send push notification to user
      operationId: sendOne
      security:
        - ServiceAuth: [ ]
      requestBody:
        required: true
        content:
          application/vnd.kaiqkt_push_v1+json:
            schema:
              $ref: '#/components/schemas/PushV1'
      responses:
        204:
          description: Push sent with successfully
    get:
      tags:
        - Notification
      description:
        Get notifications for user
      operationId: findAll
      security:
        - CustomerAuth: [ ]
      responses:
        200:
          description: Get notification history successfully
          content:
            application/vnd.kaiqkt_push_v1+json:
              schema:
                $ref: '#/components/schemas/NotificationHistoryV1'
        404:
          description: Notification history not found

components:
  schemas:
    SmsV1:
      type: object
      properties:
        recipient:
          type:
            string
        template:
          $ref: '#/components/schemas/TemplateV1'
      required:
        - recipient
        - template

    EmailV1:
      type: object
      properties:
        subject:
          type: string
        recipient:
          type: string
        template:
          $ref: '#/components/schemas/TemplateV1'
      required:
        - subject
        - recipient
        - template

    PushV1:
      type: object
      properties:
        title:
          type: string
        recipient:
          type: string
        template:
          $ref: '#/components/schemas/TemplateV1'
      required:
        - title
        - recipient
        - template

    TemplateV1:
      type: object
      properties:
        url:
          type: string
        data:
          type: object
          additionalProperties:
            type: string
      required:
        - url
        - data

    NotificationV1:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        created_at:
          type: LocalDateTime
      required:
        - title
        - body
        - created_at

    NotificationHistoryV1:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/NotificationV1'


    ErrorResponseV1:
      type: object
      properties:
        details:
          type: object

  securitySchemes:
    CustomerAuth:
      type: http
      scheme: bearer
    ServiceAuth:
      type: apiKey
      name: Authorization
      in: header